pyproject.toml
[project]
name = "hackhathon-ai-employee"
version = "0.1.0"
description = "Personal AI Employee - Bronze Tier"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "watchdog>=4.0.0",
    "google-api-python-client>=2.0.0",
    "google-auth-httplib2>=0.1.0",
    "google-auth-oauthlib>=1.0.0",
    "playwright>=1.40.0",
    "python-crontab>=3.0.0",
    "elasticsearch>=8.0.0",
    "tweepy>=4.14.0",
    "linkedin-api>=2.0.0",
    "python-dotenv>=1.0.0",
]
